# Conflicts:
#	Extinguish/Extinguish/GameObject.h
#	Extinguish/Extinguish/Main.cpp
#	Extinguish/Extinguish/Transform.cpp

+++ Extinguish/Extinguish/XTime.cpp
@@ -74,9 +74,9 @@ void XTime::Signal()
 		localStack.elapsedSignals = 0;
 	}// done
 }
-double XTime::Delta()
+float XTime::Delta()
 {
-	return localStack.deltaTime;
+	return (float)localStack.deltaTime;
 }
 double XTime::SmoothDelta()
 {
